Move Task Control Block logic into separate source / header
